examples/google_adk/translation_orchestrator_agent/pyproject.toml
[project]
name = "adk-translation-orchestrator-agent-example"
version = "0.1.0"
description = "Translation orchestrator agent example using ADK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk",
    "click>=8.1.8",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "google-genai>=1.9.0", # For the orchestrator's own LLM
    "google-adk>=1.0.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2"
    # We don't need litellm here if the orchestrator itself uses a Gemini model
    # and sub-agents handle their own model dependencies.
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
a2a-sdk = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
